+++ apps/transients/api.py
@@ -40,6 +40,6 @@ class TransientListAPIView(generics.ListAPIView):
     serializer_class = TransientSerializer
     pagination_class = TransientPagination
     filter_backends = [filters.SearchFilter, filters.OrderingFilter]
-    search_fields = ["name", "origin"]
-    ordering_fields = ["name", "origin", "ra", "decl", "created_at"]
+    search_fields = ["name", "origin", "sherlock_classification"]
+    ordering_fields = ["name", "origin", "sherlock_classification", "ra", "decl", "created_at"]
     ordering = ["name"]